# Avoid overflow/conversion errors in vectorized code.