--- a/pyproject.toml
+++ b/pyproject.toml
@@ -41,4 +41,4 @@ ignore_missing_imports = true
 
 [tool.pytest.ini_options]
 addopts = "--nbval"
-testpaths = ["test", "doc"]
+testpaths = ["docs", "test"]