--- a/test/test_decode.py
+++ b/test/test_decode.py
@@ -261,8 +261,9 @@ def test_consistent_decodes_all_values(
         np.iinfo(int_dtype).min, int(np.iinfo(int_dtype).max) + 1, dtype=int_dtype
     )
 
-    # Warning here when converting bfloat16 NaNs to float64
-    npfvals = npivals.view(dtype=npfmt).astype(np.float64)
+    with np.errstate(invalid="ignore"):
+        # Warning here when converting bfloat16 NaNs to float64
+        npfvals = npivals.view(dtype=npfmt).astype(np.float64)
 
     # Scalar version
     for i, npfval in zip(npivals, npfvals):