Some vectorized code caused overflows/conversion errors.  These did not influence correctness, as they were in non-taken elememts of an `np.where`, but it appears better to remove them.  No speed implications within timing noise.
```
GFloat scalar                  :  7908.50 nsec (25 runs at size 10000)
GFloat vectorized, numpy arrays:    62.24 nsec (25 runs at size 1000000)
GFloat vectorized, JAX JIT     :     3.52 nsec (250 runs at size 1000000)
ML_dtypes                      :     3.43 nsec (500 runs at size 1000000)
```